# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "QA-Bot",
    "description": "A question and answer app for your candidates!",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:9000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "cookieAuth": {
      "type": "apiKey",
      "in": "cookie",
      "name": "connect.sid"
    }
  },
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "body",
            "description": "User email",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "summary": "Login user",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "properties": {
                "token": {
                  "type": "string"
                },
                "user": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Login": {
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "password": {
          "type": "string"
        }
      }
    },
    "User": {
      "required": [
        "email",
        "_id"
      ],
      "properties": {
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "status": {
          "type": "string"
        }
      }
    },
...
  }
}